+++ front/src/modal/component/CreateEventModal/PermissionContainer/style.module.css
@@ -1,0 +1,7 @@
+.permission_container { display: flex; align-items: center; width: 85%; } 
+
+.permission_container > h3 { display: flex; align-items: center; position: relative; left: -10px; padding: 5px 10px; border-radius: 4px; cursor: pointer; margin-right: 0px !important; } 
+.permission_container > h3 svg { margin-right: 0px; } 
+.permission_container > h3:hover { background: #f7f7f7; } 
+
+.caret_down { margin: 0px 10px; }
